Add GitHub workflow to automatically update index.yaml
The workflow is automatically triggered on push to the repo branch, but
only if some Helm chart (tgz) file is changed.

The job will fail if an existing Helm chart is changed, or the index
update looks suspicious. In such cases, the update should be made
locally using .github/update-index.sh script.